QwenAssistant/backend/requirements.txt
# Local AI Assistant - Backend Requirements

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.23.0

# Database
sqlmodel>=0.0.8
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data Validation and Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Hardware Detection
psutil>=5.9.0
GPUtil>=1.4.0

# AI Inference
onnxruntime>=1.16.0
numpy>=1.24.0
tokenizers>=0.15.0

# Security
pycryptodome>=3.18.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.5

# Production
gunicorn>=21.0.0